refactor: reuse relay health utilities
